Commit 1.2

+++ src/Commits.java
@@ -17,8 +17,8 @@ public class Commits {
         return rezult;
     }
 
-    public int getMin() {
-        int rezult = nums[0];
+    public double getMin() {
+        double rezult = nums[0];
 
         for (int i = 1; i < nums.length; i++) {
             if (nums[i] < rezult)
